requirements.txt
Flask==3.0.0
Flask-SQLAlchemy
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF
Pillow
python-dotenv==1.0.0
Werkzeug==3.0.1
eventlet==0.35.2
SQLAlchemy
gunicorn==21.2.0
Flask-Migrate
Flask-Admin
Flask-Mail
Flask-CORS
python-jose
bcrypt
Flask-PyMongo==3.0.0
pymongo==4.6.1
dnspython==2.6.1
python-engineio==4.9.0
python-socketio==5.11.1
